+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 *Breaking*
 
 * Bumping MSRV to `1.83.0`.
+* Dropping `xan plot -Y/--add-series`, since it is now possible to select multiple series as `<y>` in  `xan plot <x>, <y>`.
 
 *Features*